finalizing chat interface
